2020.08.04 13:00 Code Operation Debug #32
TODO :
1) Init FIFO - Done
2) VLCGroup() - Done
3) UpdateHistoryElement - Done
4) addbits, getbits Problem - Done
5) Flatness Test Part Out of Bound Error - Done
6) Final Line Processing - Done
7) FIFO Overflow Problem - Done
8) SamplePredict indexing Problem - Done
9) Syntax Edited - Done
10) Shifter Structure Changed - Done
11) When is Shifter Fullness increased ?? - Done
12) AttributeError: 'numpy.int32' object has no attribute 'bit_length' - Done
13) DSC Write Index Problem - Done
14) Pixel Value Debug ... Priority!!!
15) Output Buffer Write Problem... ONLY WRITES ZERO VALUE!! - Done
16) stQp, 'd' in putbit Value Type Error - Done
17) Line Buffer Value Initialize - Done
18) Zero Padding Position Problems - Done
19) RateControl Fullness and pixelCount Problem ... ing
20) Function Call Test...
    MaxResidualSize
    PopulateOrigLine
    UseICHistory - Done
21) Fine ALL not correct variables... (vlc_var.ichSelected -> ich_var.ichSelected...)
22) Pixel Component Order Problems - Done
23) Predicted Value and Original Pixel Value Compare and Debug - Partially Done
24) BlockPrediction Error Problem
    vPos [10] hPos[1266, 1268] (MMAP -> 11)
    vPos [43] hPos[1755, 1757] (MMAP -> 7)
    vPos [72] hPos[462, 464] (MMAP -> 6)
    Needed to be corrected...

+++ init_enc_params.py
@@ -197,7 +197,7 @@ class initDscConstants:
             if diff_cond:
                 self.cpntBitDepth[i] += 1
                 # range_[i] *= 2
-        self.maxSeSize = np.zeros(4)
+        self.maxSeSize = np.zeros(4, dtype = np.int32)
         if pps.bits_per_component == 16:
             self.maxSeSize[0] = self.maxSeSize[1] = self.maxSeSize[2] = self.maxSeSize[3] = 64
         else:
